(rebased with main) fixes with refactor with NumberKeybouard, move to one uni package
