+++ app/src/main/java/kg/devcats/compose/samples/ui/navigation/SampleNavGraph.kt
@@ -16,7 +16,6 @@ import kg.devcats.compose.samples.ui.chili_sample.PreviewAutoScrollBanners
 import kg.devcats.compose.samples.ui.chili_sample.PreviewButtons
 import kg.devcats.compose.samples.ui.chili_sample.PreviewDialogs
 import kg.devcats.compose.samples.ui.chili_sample.PreviewDividers
-import kg.devcats.compose.samples.ui.chili_sample.PreviewKeyboard
 import kg.devcats.compose.samples.ui.chili_sample.PreviewPdfViewerNavGraph
 import kg.devcats.compose.samples.ui.chili_sample.TextAppearance
 import kg.devcats.compose.samples.ui.chili_sample.Toolbars
@@ -63,9 +62,6 @@ fun SamplesNavGraph(navController: NavHostController = rememberNavController()) 
         }
         composable(Screens.DividersScreen.toString()) {
             PreviewDividers { navController.navigateUp() }
-        }
-        composable(Screens.NumberKeyboardScreen.toString()) {
-            PreviewKeyboard { navController.navigateUp() }
         }
         composable(Screens.PdfViewerNavGraph.toString()) {
             PreviewPdfViewerNavGraph(navController = navController)
